Use StepNodeOptions in foundry_step_node
